Move elements outside chromedriver
