+++ lncrawl/chromedriver/chrome.py
@@ -6,11 +6,8 @@ import json
 import locale
 import logging
 import os
-from enum import Enum
 from threading import Lock, Semaphore
-from typing import List, Optional, Union
-
-from bs4 import Tag
+from typing import List, Optional
 
 from ..core.exeptions import LNException
 from ..core.soup import SoupMaker
@@ -25,13 +22,10 @@ except ImportError:
     logger.warn("`webdriver-manager` is not found")
 
 try:
-    import selenium.webdriver.support.expected_conditions as EC
     from selenium.webdriver.chrome.options import Options as ChromeOptions
     from selenium.webdriver.chrome.webdriver import WebDriver
     from selenium.webdriver.remote.command import Command
     from selenium.webdriver.remote.remote_connection import LOGGER
-    from selenium.webdriver.remote.webelement import WebElement as OriginalWebElement
-    from selenium.webdriver.support.relative_locator import RelativeBy
     from selenium.webdriver.support.wait import WebDriverWait
 
     LOGGER.setLevel(logging.ERROR)
@@ -40,11 +34,7 @@ except ImportError:
 
 
 __all__ = [
-    "By",
-    "EC",
     "Command",
-    "RelativeBy",
-    "WebElement",
     "WebDriverWait",
     "create_chrome",
     "check_if_active",
@@ -243,58 +233,3 @@ def cleanup():
 
 
 atexit.register(cleanup)
-
-
-# from selenium.webdriver.common.by import By
-class By(str, Enum):
-    ID = "id"
-    XPATH = "xpath"
-    LINK_TEXT = "link text"
-    PARTIAL_LINK_TEXT = "partial link text"
-    NAME = "name"
-    TAG_NAME = "tag name"
-    CLASS_NAME = "class name"
-    CSS_SELECTOR = "css selector"
-
-    def __str__(self) -> str:
-        return self.value
-
-
-class WebElement(OriginalWebElement):
-    def __init__(self, parent, id_):
-        super().__init__(parent, id_)
-
-    @property
-    def __soup_maker(self) -> SoupMaker:
-        if hasattr(self._parent, "soup_maker"):
-            maker = getattr(self._parent, "soup_maker")
-            if isinstance(maker, SoupMaker):
-                return maker
-        return SoupMaker()
-
-    def inner_html(self) -> str:
-        return self.get_attribute("innerHTML")
-
-    def outer_html(self) -> str:
-        return self.get_attribute("outerHTML")
-
-    def as_tag(self) -> Tag:
-        html = self.outer_html()
-        if not hasattr(self, "_tag") or self._html != html:
-            self._html = html
-            self._tag = self.__soup_maker.make_tag(self._tag)
-        return self._tag
-
-    def find_all(
-        self, selector: str, by: Union[By, RelativeBy] = By.CSS_SELECTOR
-    ) -> List["WebElement"]:
-        if isinstance(by, By):
-            by = str(by)
-        self.find_elements(by, selector)
-
-    def find(
-        self, selector: str, by: Union[By, RelativeBy] = By.CSS_SELECTOR
-    ) -> "WebElement":
-        if isinstance(by, By):
-            by = str(by)
-        self.find_element(by, selector)
